requirements.txt
opencv-python
mediapipe
numpy
torchsummary
Pillow